correct y axis and set custom bounding box
the y axis is inverted and needs to be set to true. currently when you control the bed, if you press the up arrow on the control tab it will move the bed towards you. this isnt intuitive as you'd expect the bed to be moved away from you as normally you are looking at the bed from the LCD side (front) in the webcam.

also the y axis is allowed to technically move to -4, setting a custom bound box prevents errors for plugins such as octolapse who expect the printer to be within the bed WxHxD or custom bound box.